+++ src/main/java/com/study/realworld/configuration/config/WebConfig.java
@@ -1,4 +1,4 @@
-package com.study.realworld.config;
+package com.study.realworld.configuration.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.bind.annotation.RestController;
