+++ src/main/java/com/study/realworld/configuration/util/StringUtils.java
@@ -1,0 +1,7 @@
+package com.study.realworld.configuration.util;
+
+public class StringUtils {
+    public static boolean isNotNullAndNotEmpty(final String str) {
+        return str != null && !str.isEmpty();
+    }
+}
