+++ src/main/java/com/study/realworld/domain/user/dto/UserUpdateRequestDTO.java
@@ -1,4 +1,4 @@
-package com.study.realworld.dto;
+package com.study.realworld.domain.user.dto;
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.annotation.JsonTypeName;
