--- src/main/java/com/study/realworld/util/StringUtils.java
@@ -1,7 +1,0 @@
-package com.study.realworld.util;
-
-public class StringUtils {
-    public static boolean isEmptyOrNull(final String str) {
-        return str == null || str.isEmpty();
-    }
-}
